feat(Booking): add Booking.by_date scope
Querying by value date is a common use case when working with Booking
objects.

This patch introduces a .by_date scope accepting single dates or
two parameters for start and/or end date.

This scope overlapps with the .by_value_period and .by_value_date methods.
They will be deprecated in a further release.