spec: improve account factory
The account factory does not create new Account entries for the same
code, but returns existing ones. This is good for performance purposes
but also simplifies creating booking instances for designated accounts.
It did only create a single instance when called without an explicit
code. This is not the intended behaviour.

This patch assigns a factored account instance a random code.