feat(Account): implement Account.find_by_tag
We support looking up Accounts by tags when the acts-as-taggable-on gem
is used. Looking up by tags might lead to multiple findings, which is
not always what we want. When needing a single instance, as when
assigning to a Booking, you may use the `.find_by_tag` method which
raises an `Account::AmbiguousTag` error if multiple Accounts are
candidates.

Thanks go to luxflux for implementing this while working on API support
for Bookyt: https://github.com/huerlisi/bookyt/pull/82